requirements.txt
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.15.0
pandas==0.23.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
